Rename desync -> resync
